+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/repositories/VideosRepository.kt
@@ -4,6 +4,6 @@ import com.denisyordanp.mymoviecatalogue.schemas.ui.Video
 import kotlinx.coroutines.flow.Flow
 
 interface VideosRepository {
-    suspend fun reloadVideos(movieId: Int)
-    fun getVideos(movieId: Int): Flow<List<Video>>
+    suspend fun reloadVideos(movieId: Long)
+    fun getVideos(movieId: Long): Flow<List<Video>>
 }
