+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/schemas/ui/Movie.kt
@@ -1,11 +1,11 @@
 package com.denisyordanp.mymoviecatalogue.schemas.ui
 
 data class Movie(
-    val id: Int,
+    val id: Long,
     val overview: String,
     val posterPath: String,
     val releaseDate: String,
     val title: String,
     val voteAverage: Double,
-    val voteCount: Int
+    val voteCount: Long
 )
