+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/usecase/GetMovies.kt
@@ -4,5 +4,5 @@ import com.denisyordanp.mymoviecatalogue.schemas.ui.Movie
 import kotlinx.coroutines.flow.Flow
 
 fun interface GetMovies {
-    operator fun invoke(genreId: Int, isForce: Boolean): Flow<List<Movie>>
+    operator fun invoke(genreId: Long, isForce: Boolean): Flow<List<Movie>>
 }
